ESLint and organization updates
