+++ src/client/utils/puzzle-functions/makeAllPeers.ts
@@ -6,14 +6,17 @@ import { allSquareIds } from './allSquareIdsAndPuzzleVals';
 
 /** makeAllPeers
  *
- * Two squares are peers if they can't simultaneously hold the same number in a Sudoku puzzle. This includes
- * any two numbers in the same row, column, or box (set of 9 squares) are peers.
- * makeAllPeers generates an allPeers object that holds key:value pairs where the key is a SquareId string (e.g. 'A1')
- * and the value is a set of all of the squareId string peers (e.g. 'A2', 'A3',...) of that key.
- * This set will be used in the eventual population of the squares. An array of all boxes is also returned for use
- * in the Sudoku grid.
+ * Two squares are peers if they can't simultaneously hold the same number in a Sudoku puzzle. This
+ * includes any two numbers in the same row, column, or box (set of 9 squares) are peers.
+ * makeAllPeers generates an allPeers object that holds key:value pairs where the key is a SquareId
+ * string (e.g. 'A1') and the value is a set of all of the squareId string peers
+ * (e.g. 'A2', 'A3',...) of that key.
+ * This set will be used in the styling of the squares. Other groupings of squares is also returned
+ * for use in the Sudoku grid.
  *
- * @returns an allPeers object and a boxes array that holds 9 sets of the squares that form the sudoku grid
+ * @returns an allPeers object, a boxes array that holds 9 sets of the squares that form the
+ * sudoku grid, a rows array that holds 9 sets of squareIds in each row, and a column array that
+ * holds 9 sets of squareIds in each column.
  */
 
 const makeAllPeers = (): {
@@ -46,7 +49,8 @@ const makeAllPeers = (): {
     // e.g. boxes[0] = { 'A1', 'A2', 'A3', 'B1', 'B2', 'B3', 'C1', 'C2', 'C3' }
     // Each index divided by 3 and then modulo 3 is the same for each unit box
     const howFarFromDivisionByThree = Math.floor(i / 3) % 3;
-    // Each of the 81 SquareIds are grouped in one of three sections three sections (aka 9 square row), 0, 1, and 2
+    // Each of the 81 SquareIds are grouped in one of three sections three sections
+    // (aka 9 square row), 0, 1, and 2
     const section = Math.floor(i / 27);
     boxes[section * 3 + howFarFromDivisionByThree].add(allSquareIds[i]);
 
@@ -56,8 +60,8 @@ const makeAllPeers = (): {
   // Add every grouping of related squareId's to an array
   const allUnits = rows.concat(cols).concat(boxes);
 
-  // For every squareId, find every grouping it's a part of and add every other squareId (peer) in that grouping
-  // to the set in the allPeers object at that squareId's key
+  // For every squareId, find every grouping it's a part of and add every other squareId (peer)
+  // in that grouping to the set in the allPeers object at that squareId's key
   for (const squareId of allSquareIds) {
     for (const peerUnit of allUnits) {
       if (peerUnit.has(squareId)) {
