+++ src/client/utils/puzzle-functions/updateSquaresDuplicates.ts
@@ -14,9 +14,9 @@ import { allPeers } from './makeAllPeers';
 
 /** updateFilledSquaresDuplicates
  *
- * Iterates over a filledSquares object and updates the value of each filledSquare's duplicate property so that
- * it's accurate based on its filledSquare and pencilSquare peers. It's assumed that the filledSquares parameter
- * provided has already been deep copied
+ * Iterates over a filledSquares object and updates the value of each filledSquare's duplicate
+ * property so that it's accurate based on its filledSquare and pencilSquare peers. It's assumed
+ * that the filledSquares parameter provided has already been deep copied
  *
  * @param filledSquares - FilledSquares object
  * @param pencilSquares - PencilSquares object
@@ -27,7 +27,8 @@ export const updateFilledSquaresDuplicates = (
 ): void => {
   // Iterate over every filledSquare in filledSquares
   const squareIds = Object.keys(filledSquares).filter((key) => key !== 'size') as SquareId[];
-  // Update each filledSquare's duplicate status based on its peers filledSquare and pencilSquare values
+  // Update each filledSquare's duplicate status based on its peers filledSquare and
+  // pencilSquare values
   for (const squareId of squareIds) {
     const square = filledSquares[squareId] as FilledSquare;
     square.duplicate = false;
@@ -40,9 +41,9 @@ export const updateFilledSquaresDuplicates = (
 
 /** updatePencilSquaresDuplicates
  *
- * Iterates over a pencilSquares object and updates the value of each pencilSquare's duplicate property so that
- * it's accurate based on it's filledSquare peers. It's assumed that the pencilSquares parameter provided has
- * already been deep copied
+ * Iterates over a pencilSquares object and updates the value of each pencilSquare's duplicate
+ * property so that it's accurate based on it's filledSquare peers. It's assumed that the
+ * pencilSquares parameter provided has already been deep copied
  *
  * @param filledSquares - FilledSquares object
  * @param pencilSquares - PencilSquares object
